Refactored positions, set and edges
In this commit, I made this three arrays (positions, set and edges) from global variables to variables of the class/object Maze. Also, class Icon now needs to receive maze because it uses positions to check if the player won and positions is now a variable of maze. Probably in the future, the checkWinner() funtion will be in the Maze class and not in the Icon.